+++ components/blocks-renderer.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 import { Feature } from "./blocks/feature";
 import { Cards } from "./blocks/cards";
+import { Carousel } from "./blocks/carousel";
 import { Accordion } from "./blocks/accordion";
 import { Embed } from "./blocks/embed";
 import { TailwindFeature } from "./blocks/tailwind-feature";
 import { TailwindCards } from "./blocks/tailwind-cards";
 import { EventSchedule } from "./blocks/event-schedule";
 import { EventTimeline } from "./blocks/event-timeline";
-import { tinaField } from "tinacms/dist/react";
+import { Video } from "./blocks/video";
 
 export const Blocks = (props: any) => {
   return (
@@ -18,7 +19,6 @@ export const Blocks = (props: any) => {
               case "PageBlocksFeature":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Feature data={block} />
@@ -27,16 +27,22 @@ export const Blocks = (props: any) => {
               case "PageBlocksCards":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Cards data={block} />
                   </div>
                 );
+              case "PageBlocksCarousel":
+                return (
+                  <div
+                    key={i + block.__typename}
+                  >
+                    <Carousel data={block} />
+                  </div>
+                );
               case "PageBlocksAccordion":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Accordion data={block} />
@@ -45,7 +51,6 @@ export const Blocks = (props: any) => {
               case "PageBlocksEmbed":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Embed data={block} />
@@ -54,7 +59,6 @@ export const Blocks = (props: any) => {
               case "PageBlocksTailwindFeature":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <TailwindFeature data={block} />
@@ -63,7 +67,6 @@ export const Blocks = (props: any) => {
               case "PageBlocksTailwindCards":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <TailwindCards data={block} />
@@ -72,7 +75,6 @@ export const Blocks = (props: any) => {
               case "PageBlocksEventSchedule":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <EventSchedule data={block} events={props.events as any} />
@@ -81,10 +83,18 @@ export const Blocks = (props: any) => {
               case "PageBlocksEventTimeline":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <EventTimeline data={block} events={props.events} />
+                  </div>
+                );
+              case "PageBlocksVideo":
+                return (
+                  <div
+                    data-tinafield={`blocks.${i}`}
+                    key={i + block.__typename}
+                  >
+                    <Video data={block} />
                   </div>
                 );
               default:
@@ -93,7 +103,6 @@ export const Blocks = (props: any) => {
               case "GlobalBlocksFeature":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Feature data={block} />
@@ -102,7 +111,6 @@ export const Blocks = (props: any) => {
               case "GlobalBlocksCards":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Cards data={block} />
@@ -111,7 +119,6 @@ export const Blocks = (props: any) => {
               case "GlobalBlocksEmbed":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <Embed data={block} />
@@ -120,7 +127,6 @@ export const Blocks = (props: any) => {
               case "GlobalBlocksTailwindFeature":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <TailwindFeature data={block} />
@@ -129,7 +135,6 @@ export const Blocks = (props: any) => {
               case "GlobalBlocksTailwindCards":
                 return (
                   <div
-                    data-tina-field={tinaField(block)}
                     key={i + block.__typename}
                   >
                     <TailwindCards data={block} />